requirement.txt
torch
torchvision
opencv-python
matplotlib
Pillow
gdown
